+++ puma/hlplots/__init__.py
@@ -2,6 +2,7 @@
 from __future__ import annotations
 
 from puma.hlplots.configs import PlotConfig
+from puma.hlplots.aux_results import AuxResults
 from puma.hlplots.results import Results
 from puma.hlplots.tagger import Tagger
 from puma.hlplots.yutils import (
@@ -14,9 +15,11 @@ from puma.hlplots.yutils import (
 __all__ = [
     "Results",
     "Tagger",
+    "AuxResults",
     "PlotConfig",
     "get_included_taggers",
     "get_plot_kwargs",
     "get_signals",
     "select_configs",
 ]
+