+++ puma/tests/utils/test_generate.py
@@ -8,7 +8,11 @@ import numpy as np
 from numpy.lib.recfunctions import structured_to_unstructured as s2u
 
 from puma.utils import logger, set_log_level
-from puma.utils.generate import get_dummy_2_taggers, get_dummy_multiclass_scores
+from puma.utils.generate import (
+    get_dummy_2_taggers,
+    get_dummy_multiclass_scores,
+    get_dummy_tagger_aux,
+)
 
 set_log_level(logger, "DEBUG")
 
@@ -54,3 +58,25 @@ class GetDummy2TaggersTestCase(unittest.TestCase):
         """Check correct amount of columns when using pt as well."""
         df_gen = get_dummy_2_taggers(add_pt=True)
         self.assertEqual(len(df_gen.columns.values), 9)
+
+
+class GetDummyTaggerAuxTestCase(unittest.TestCase):
+    """Test case for get_dummy_tagger_aux function."""
+
+    def test_size(self):
+        """Check that correct size is returned."""
+        file = get_dummy_tagger_aux(size=10)
+        # we expect here 9 entries, since 10 is not dividable by 3, and the function
+        # returns 3 classes with the same amount of stats per class
+        self.assertEqual(file["jets"].shape[0], 9)
+
+    def test_ntracks(self):
+        """Check correct amount of tracks."""
+        file = get_dummy_tagger_aux(n_tracks=10)
+        self.assertEqual(file["tracks"].shape[1], 10)
+
+    def test_columns(self):
+        """Check correct amount of columns for jets and tracks."""
+        file = get_dummy_tagger_aux()
+        self.assertEqual(len(file["jets"][0]), 7)
+        self.assertEqual(len(file["tracks"][0][0]), 2)